simple-travel-backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
langchain-core==0.1.52
langchain-groq==0.1.4
pandas==2.2.2
python-dotenv==1.0.1
